requirements.txt
PyMuPDF
python-dotenv
google-generativeai
scikit-learn
numpy
tiktoken 